worldmap-pan-zoom-mouse
Implemented logic to navigate, zoomm or pan across worldmap. Created usePan hook which can be used to use mouse drag gestures to move across the graphs. Created NavControls component which is a set of floating buttons UI for navigation using zoom or scale. The hook and component are designed for the Graph component but some adjustments could be required in order for them to be implemented for graphs other worldmap.